# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "influxdb practice",
        "title": "Ron influxdb swagger demo",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/influx/v1/hello": {
            "get": {
                "tags": [
                    "Hello"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/influx/v1/temperature": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Temperature"
                ],
                "parameters": [
                    {
                        "description": "location deviceId temperature",
                        "name": "temperatureRecord",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TemperatureRecord"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/influx/v1/temperature/avg/{location}/{deviceId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Temperature"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "deviceId",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "location",
                        "name": "location",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.TemperatureRecord": {
            "type": "object",
            "properties": {
                "deviceId": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "temperature": {
                    "type": "number"
                }
            }
        }
    }
}